перенёс имеющиеся методы в api сервис
